이윤서: 아이템 생성 시 회전 및 드롭

+++ GameEngineContents/Monster.h
@@ -166,8 +166,8 @@ private:
 	float CanAttTime_;
 	float AttEndTime_;
 
-	float BottomPos_;
-	float LeftRightPos_;
+	int BottomPos_;
+	int LeftRightPos_;
 	float4 DownPower_;
 
 	float4 MoveDir_;
